RoleAction: Removed Fetch One Role Action
Overloaded Fetch Role Action with optional role parameter

+++ src/actions/RolesAction.js
@@ -8,47 +8,30 @@ export let FETCH_PERMISSION_ACTION = 'FETCH_PERMISSION_ACTION';
 export let UPDATE_ROLE_ACTION = 'UPDATE_ROLE_ACTION';
 export let DELETE_ROLE_ACTION = 'DELETE_ROLE_ACTION';
 
-// TODO: Fetch Role Details from API
-export function actionFetchRoles() {
-  var url = BASE_URL + "roles";
-  console.log("Fetching roles from API.");
-  const request = axios.get(url);
-  console.log("Fetch roles request response:", request);
-  return {
-    type: FETCH_ROLE_ACTION,
-    payload: request
-  };
-}
+// Target URL
+var url = BASE_URL + "roles";
 
-// TODO: Fetch single Role Details
-export function actionFetchOneRole(role) {
-    var url = BASE_URL + `roles/${role}`;
-    console.log("Fetching roles from API for role", role);
+// TODO: Fetch Role Details from API
+export function actionFetchRoles(role) {
+    actionType = FETCH_ROLE_ACTION;
+    if (typeof role !== 'undefined') {
+        url += `/${role}`;
+        actionType = FETCH_ONE_ROLE_ACTION;
+    }
+    console.log("Fetching roles from API.");
     const request = axios.get(url);
     console.log("Fetch roles request response:", request);
     return {
-      type: FETCH_ONE_ROLE_ACTION,
-      payload: request
+        type: actionType,
+        payload: request
     };
 }
 
-// TODO: Get the list of Components available
-// export function actionFetchComponents() {
-//     var url = BASE_URL + "components";
-//     console.log("Fetching components from API.");
-//     const request = axios.get(url);
-//     console.log("Fetch components request response:", request);
-//     return {
-//         type: FETCH_COMPONENT_ACTION,
-//         payload: request
-//     };
-// }
-
 // TODO: Get the list of Permissions available
 export function actionFetchPermissions() {
-    var url = BASE_URL + "permissions";
+    var purl = BASE_URL + "permissions";
     console.log("Fetching permissions from API.");
-    const request = axios.get(url);
+    const request = axios.get(purl);
     console.log("Fetch permissions request response:", request);
     return {
         type: FETCH_PERMISSION_ACTION,
@@ -58,24 +41,25 @@ export function actionFetchPermissions() {
 
 // TODO: Send role data to API
 export function actionUpdateRoles(data) {
-    var url = BASE_URL + "roles";
     console.log("Sending roles to API.", data);
     const request = axios.post(url, data);
     console.log("Update roles request response:", request);
     return {
-      type: UPDATE_ROLE_ACTION,
-      payload: request
+        type: UPDATE_ROLE_ACTION,
+        payload: request
     };
 }
 
 // TODO: Delete role data
 export function actionDeleteRoles(role) {
-    var url = BASE_URL + `roles/${role}`;
-    console.log("Deleting roles to API.", role);
-    const request = axios.delete(url);
-    console.log("Delete roles request response:", request);
-    return {
-      type: DELETE_ROLE_ACTION,
-      payload: request
-    };
+    if (typeof role !== 'undefined') {
+        url += `/${role}`;
+        console.log("Deleting roles to API.", role);
+        const request = axios.delete(url);
+        console.log("Delete roles request response:", request);
+        return {
+            type: DELETE_ROLE_ACTION,
+            payload: request
+        };
+    }
 }
